Fix istringstream read for readlines
When using operator>> on a istringstream, there is a chance that it will
silently fail if the contents of the buffer have all already been read.
This is not a major issue in most scenarios. However, in a situation like
this, the while loop will read all of the lines in, get to the last
line, and then (I believe, I'm not 100% sure on this) read the EOF but
still continue into the loop. As a result, using operator>> with 'line'
during the last iteration will fail and the contents of 'name', 'mass',
and 'radius' will stay the same. Essentially, without the if-statement
the last line will get printed twice. Since this code is used as an
example for the robosim individual assignment, this is pertinent to fix
before it gets used again.

individual assignment